Corrected some string related mismatches                          [odin]
In 'heapstring.h', 'strncasecmp' is mapped to 'lstrncmpiA', but the
latter used 'int' for the length and not 'size_t'. According to the C++
language documentation, 'size_t' is an 'unsigned int'.

git-svn-id: http://svn.netlabs.org/repos/odin32/branches/swt@22123 3ae8c665-79e0-dc11-ae90-00e081727c95

+++ include/win/winbase.h
@@ -2270,8 +2270,8 @@ INT       WINAPI lstrcmpW(LPCWSTR,LPCWSTR);
 INT       WINAPI lstrcmpiA(LPCSTR,LPCSTR);
 INT       WINAPI lstrcmpiW(LPCWSTR,LPCWSTR);
 #define     lstrcmpi WINELIB_NAME_AW(lstrcmpi)
-int       WINAPI lstrncmpiA(LPCSTR, LPCSTR, int);
-int       WINAPI lstrncmpiW(LPCWSTR, LPCWSTR, int);
+int       WINAPI lstrncmpiA(LPCSTR, LPCSTR, size_t);
+int       WINAPI lstrncmpiW(LPCWSTR, LPCWSTR, size_t);
 #define     lstrncmpi WINELIB_NAME_AW(lstrncmpi)
 
 VOID        WINAPI SetLastError(DWORD);
